Added the first Version of KI-AI to the Repository.
Currently has a Search tool with partially working text responses to the searched things. It also has the foundations of a chat function. Current plans is to use buttons for conversation until language models such as OPEN AI can be implemented. There is no colour, and there is only a grid for the text to go on.